Fixes #64
Polished current location show on map.
